Move TE bits around a bit
