+++ .gitignore
@@ -234,3 +234,4 @@ AkavacheSqliteLinkerOverride.cs
 NuGetBuild
 WiX.Toolset.DummyFile.txt
 nunit-UnitTests.xml
+GitHubVS.sln.DotSettings
