+++ src/GitHub.VisualStudio/Base/TeamExplorerServiceHolder.cs
@@ -156,7 +156,7 @@ namespace GitHub.VisualStudio.Base
                 var page = PageService;
                 if (page == null)
                     return null;
-                return page.GetSection(new Guid(TeamExplorerHome.GitHubHomeSection.GitHubHomeSectionId)) as IGitAwareItem;
+                return page.GetSection(new Guid(TeamExplorer.Home.GitHubHomeSection.GitHubHomeSectionId)) as IGitAwareItem;
             }
         }
