+++ src/GitHub.VisualStudio/GitHub.VisualStudio.csproj
@@ -244,17 +244,17 @@
     <Compile Include="PkgCmdID.cs" />
     <Compile Include="Services\UIProvider.cs" />
     <Compile Include="SimpleJson.cs" />
-    <Compile Include="TeamExplorerConnect\GitHubConnectSection1.cs" />
-    <Compile Include="TeamExplorerConnect\GitHubConnectSection0.cs" />
-    <Compile Include="TeamExplorerConnect\GitHubConnectSection.cs" />
-    <Compile Include="TeamExplorerConnect\GitHubInvitationSection.cs" />
-    <Compile Include="TeamExplorerHome\GitHubHomeSection.cs" />
-    <Compile Include="TeamExplorerHome\GraphsNavigationItem.cs" />
-    <Compile Include="TeamExplorerHome\PullRequestsNavigationItem.cs" />
-    <Compile Include="TeamExplorerHome\IssuesNavigationItem.cs" />
-    <Compile Include="TeamExplorerHome\WikiNavigationItem.cs" />
-    <Compile Include="TeamExplorerHome\PulseNavigationItem.cs" />
-    <Compile Include="TeamExplorerSync\GitHubPublishSection.cs" />
+    <Compile Include="TeamExplorer\Connect\GitHubConnectSection1.cs" />
+    <Compile Include="TeamExplorer\Connect\GitHubConnectSection0.cs" />
+    <Compile Include="TeamExplorer\Connect\GitHubConnectSection.cs" />
+    <Compile Include="TeamExplorer\Connect\GitHubInvitationSection.cs" />
+    <Compile Include="TeamExplorer\Home\GitHubHomeSection.cs" />
+    <Compile Include="TeamExplorer\Home\GraphsNavigationItem.cs" />
+    <Compile Include="TeamExplorer\Home\PullRequestsNavigationItem.cs" />
+    <Compile Include="TeamExplorer\Home\IssuesNavigationItem.cs" />
+    <Compile Include="TeamExplorer\Home\WikiNavigationItem.cs" />
+    <Compile Include="TeamExplorer\Home\PulseNavigationItem.cs" />
+    <Compile Include="TeamExplorer\Sync\GitHubPublishSection.cs" />
     <Compile Include="UI\DrawingExtensions.cs" />
     <Compile Include="UI\Views\Controls\RepositoryCloneControl.xaml.cs">
       <DependentUpon>RepositoryCloneControl.xaml</DependentUpon>
