+++ src/GitHub.VisualStudio/TeamExplorer/Connect/GitHubConnectSection1.cs
@@ -4,7 +4,7 @@ using GitHub.Services;
 using Microsoft.TeamFoundation.Controls;
 using System.ComponentModel.Composition;
 
-namespace GitHub.VisualStudio.TeamExplorerConnect
+namespace GitHub.VisualStudio.TeamExplorer.Connect
 {
     [TeamExplorerSection(GitHubConnectSection1Id, TeamExplorerPageIds.Connect, 10)]
     [PartCreationPolicy(CreationPolicy.NonShared)]
